requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
playwright
beautifulsoup4
langgraph
langchain-core
langchain-community
langchain-google-genai
pandas
numpy
scikit-learn
scipy
openpyxl
pypdf
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchaudio
torchvision
openai-whisper
ffmpeg-python
